+++ lib/src/subtypes/guilds/guild_manager.dart
@@ -1,6 +1,6 @@
 import 'package:collection/collection.dart';
 
-import '../../randpg_exceptions.dart';
+import '../../exceptions/not_found_exceptions.dart';
 import 'artisans_guild/artisans_guild.dart';
 import 'cult/cult.dart';
 import 'druid_circle/druid_circle.dart';