+++ lib/src/subtypes/worlds/world_settings_manager.dart
@@ -1,6 +1,6 @@
 import 'package:collection/collection.dart';
 
-import '../../randpg_exceptions.dart';
+import '../../exceptions/not_found_exceptions.dart';
 import 'default_settings/default_world_settings.dart';
 import 'world_settings.dart';